feat: cliente logic and endpoint

+++ ApplicativoTalhonario/Bussines/Interfaces/Repository/IClientRepository.cs
@@ -8,8 +8,8 @@ public interface IClientRepository
 {
     public void saveClient(Client client);
     
-    public Client findById(int id);
+    public Client findById(long id);
     
     
-    public void deleteClient(int id);
+    public void deleteClient(long id);
 }
